topsis_web/requirements.txt
flask
pandas
numpy
python-dotenv
